backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
backoff==2.2.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.129.0
googleapis-common-protos==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==1.3.4
langchain-core==1.2.18
langfuse==3.14.5
langgraph==1.1.0
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.10
langsmith==0.7.16
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.26.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
protobuf==6.33.5
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.8.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.52.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==1.17.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
